project-management-agent/requirements.txt
# Project Management Agent - Dependencies

# Core AI
anthropic>=0.18.0,<1.0

# CLI and UI
rich>=13.7.0,<14.0
typer>=0.9.0,<1.0

# Templates and Configuration
jinja2>=3.1.0,<4.0
pydantic>=2.5.0,<3.0
python-dotenv>=1.0.0,<2.0
pyyaml>=6.0,<7.0

# Logging
loguru>=0.7.0,<1.0

# Utilities
requests>=2.31.0,<3.0
tqdm>=4.65.0,<5.0

# Database
sqlalchemy>=2.0.0,<3.0

# Optional: RAG System (install if needed)
chromadb>=0.4.20,<0.5
sentence-transformers>=2.2.0,<3.0

# Optional: Web Interface (install if needed)  
streamlit>=1.30.0,<2.0
fastapi>=0.100.0,<1.0

# Optional: Document Processing (install if needed)
python-docx>=1.1.0,<2.0
openpyxl>=3.1.0,<4.0
pymupdf>=1.23.0,<2.0
